checkout

+++ code/BUPT_histogram_modeling.m
@@ -1,11 +1,12 @@
+% image created by Zhengxiao Wu
 function modeled_img = BUPT_histogram_modeling(img_A, img_B)
     % Compute the histogram values of image A and B
     hist_A = BUPT_histogram(img_A);
     hist_B = BUPT_histogram(img_B);
 
     % Compute the cumulative distribution functions of the histograms
-    cdf_A = cumsum(hist_A) / numel(img_A);
-    cdf_B = cumsum(hist_B) / numel(img_B);
+    cdf_A = double(cumsum(hist_A) / numel(img_A));
+    cdf_B = double(cumsum(hist_B) / numel(img_B));
 
     % Compute the mapping index
     mapping_index = zeros(1, 256);
